+++ scripts/build-themes.js
@@ -29,6 +29,9 @@ const THEMES_PACKAGES = [
   '@adv-ui/mt-theme'
 ]
 
+const cwd = process.cwd()
+const {CI, NODE_AUTH_TOKEN} = process.env
+
 const writeFile = (path, body) =>
   fse
     .outputFile(path, body)
@@ -43,28 +46,29 @@ const writeFile = (path, body) =>
 const checkFileExists = path => fse.pathExists(path)
 
 const getThemesList = async () => {
+  console.log('[sui-studio] Getting themes list...')
   const files = await fs.readdir(path.join(__dirname, '..', 'themes'))
   return files
     .filter(file => !file.startsWith('_'))
     .map(file => file.split('.')[0])
 }
 
-const installThemesPkgs = () =>
-  exec(`npm i ${THEMES_PACKAGES.join(' ')} ${INSTALL_FLAGS.join(' ')}`, {
-    cwd: process.cwd()
+const installThemesPkgs = () => {
+  console.log('[sui-studio] Installing themes packages...')
+  return exec(`npm i ${THEMES_PACKAGES.join(' ')} ${INSTALL_FLAGS.join(' ')}`, {
+    cwd
   })
+}
 
 const writeThemesInDemoFolders = async themes => {
+  console.log('[sui-studio] Writing themes files on demos...')
   await exec('rm -Rf components/**/**/demo/themes', {
-    cwd: process.cwd()
+    cwd
   })
 
   const paths = await globby(
-    [
-      path.join(process.cwd(), 'components', '**', '**', 'demo'),
-      '!**/node_modules/**'
-    ],
-    {onlyDirectories: true, cwd: process.cwd()}
+    [path.join(cwd, 'components', '**', '**', 'demo'), '!**/node_modules/**'],
+    {onlyDirectories: true, cwd}
   )
 
   paths
@@ -85,12 +89,19 @@ ${hasDemoStyles ? `@import '../index.scss';` : ''}
           )
         )
       } catch (e) {
-        console.log('Err:', e)
+        console.log('Error:', e)
       }
     })
 }
 
 ;(async () => {
+  if (CI && !NODE_AUTH_TOKEN) {
+    console.log(
+      '[sui-studio] Skipping themes installation as NODE_AUTH_TOKEN is not available'
+    )
+    return
+  }
+
   const themes = await getThemesList()
   await installThemesPkgs()
   await writeThemesInDemoFolders(themes)